Changed format of code to es6 modules. Formatted all the code neatly.
