+++ public/index.html
@@ -1,67 +1,73 @@
 <!-- This html file is what the browser shows when we run index.js on port 5000. Therefore our display of our game-->
 <!DOCTYPE html>
 <html>
-    <head>
-        <title>Aftermath</title>
 
-        <!-- SCRIPTS-->
-        <!--Loads three.js into our game-->
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.js"> </script>
-        <!--Loads orbit controls around player-->
-        <script src="https://cdn.rawgit.com/mrdoob/three.js/master/examples/js/controls/OrbitControls.js"></script>
-        <!--Loads gltf files (our blender files)-->
-        <script src="https://cdn.rawgit.com/mrdoob/three.js/master/examples/js/loaders/GLTFLoader.js"></script>
-        <!--Pointer lock controls-->
-        
-       
+<head>
+  <title>Aftermath</title>
 
-
-        <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-        <style>
-            body { margin: 0; height: 100%; width: 100%; overflow: hidden}
-        </style>
-
-      </head>
-
-    <body>
-        <canvas id="canvas"></canvas>
-        <!--MOVEMENT-->
-        <script src="./important/THREEx.KeyboardState.js" ></script>
-        <script src="./important/FirstPersonControls.js" ></script>
-      <!--  <script src="./important/PointerLockControls.js" ></script>-->
-        <script src="./js/CharacterControls.js" ></script>
-        <!--========================================-->
-        <script src="https://threejs.org/examples/js/controls/PointerLockControls.js"></script>
-
-        <!--Lighting models-->
-        <script src="./js/sceneSubjects/lighting/GeneralLights.js" ></script>
-
-        <!--add models to scene-->
-        <!--Subject (model) scripts-->
-
-        <!--This is our main blender house-->
-        <script src="./js/sceneSubjects/House.js" ></script>
-        <!--Our character model file (not complete due to loading issues with fbx)-->
-        <script src="./js/sceneSubjects/characters/MainChar.js"  ></script>
-
-        <!--Just a demo for you to see how this works-->
-        <script src="./js/sceneSubjects/objects/SceneSubject.js" ></script>
-
-        <!--Block to test raycasting -->
-        <script src="./js/sceneSubjects/characters/TestBlock.js" ></script>
+  <!-- SCRIPTS-->
+  <!--Loads three.js into our game-->
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.js"> </script>
+  <!--Loads orbit controls around player-->
+  <script src="https://cdn.rawgit.com/mrdoob/three.js/master/examples/js/controls/OrbitControls.js"></script>
+  <!--Loads gltf files (our blender files)-->
+  <script src="https://cdn.rawgit.com/mrdoob/three.js/master/examples/js/loaders/GLTFLoader.js"></script>
+  <script src="https://rawgit.com/mrdoob/three.js/master/examples/js/loaders/FBXLoader.js"></script>
+  <!--Pointer lock controls-->
 
 
-        <!--MANAGERS-->
-        <!--Load the scene manager-->
 
-        <script src="./js/Overlay/Overlay.js" ></script>  
-        <script src="./js/EntityManager.js" ></script>  
-        <script src="./js/SceneSubjects/Menu/PauseMenu.js" ></script>  
 
-       <script src="./js/SceneManager.js" ></script>  
+  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
+  <style>
+    body {
+      margin: 0;
+      height: 100%;
+      width: 100%;
+      overflow: hidden
+    }
+  </style>
 
-            <script src = "./js/Time.js" ></script>
-        <!--Load our main.js function-->
-        <script type="module" src="./js/main2.js"></script>
-    </body>
+</head>
+
+<body>
+  <canvas id="canvas"></canvas>
+  <!--MOVEMENT-->
+  <script src="./important/THREEx.KeyboardState.js"></script>
+  <script src="./important/FirstPersonControls.js"></script>
+  <!--  <script src="./important/PointerLockControls.js" ></script>-->
+  <script src="./js/managers/KeyboardManager.js" type='module'></script>
+  <script src="./js/managers/CharacterControls.js" type='module'></script>
+  <!--========================================-->
+
+  <!--Lighting models-->
+  <script src="./js/sceneSubjects/lighting/GeneralLights.js" type='module'></script>
+
+  <!--add models to scene-->
+  <!--Subject (model) scripts-->
+
+  <!--This is our main blender house-->
+  <script src="./js/sceneSubjects/House.js" type='module'></script>
+  <!--Our character model file (not complete due to loading issues with fbx)-->
+  <script src="./js/sceneSubjects/characters/MainChar.js" type='module'></script>
+
+  <!--Just a demo for you to see how this works-->
+  <script src="./js/sceneSubjects/objects/SceneSubject.js" type='module'></script>
+
+  <!--Block to test raycasting -->
+  <script src="./js/sceneSubjects/characters/TestBlock.js" type='module'></script>
+
+  <!--MANAGERS-->
+  <script src="./js/Overlay/Overlay.js" type='module'></script>
+  <!--Load the Entity manager-->
+  <script src="./js/managers/EntityManager.js" type='module'></script>
+  <script src="./js/SceneSubjects/Menu/PauseMenu.js" type='module'></script>
+  <!--Load the scene manager-->
+  <script src="./js/managers/SceneManager.js" type='module'></script>
+
+  <script src="./js/Time.js" type='module'></script>
+  <!--Load our main.js function-->
+  <script type="module" src="./js/main2.js"></script>
+</body>
+
 </html>
