+++ public/js/A_notPartOfGameJustForChecking/check.js
