+++ public/js/SceneSubjects/House.js
@@ -1,4 +1,4 @@
-class House extends THREE.Object3D{
+export class House extends THREE.Object3D {
   constructor() {
 
     super();
@@ -18,18 +18,18 @@ class House extends THREE.Object3D{
       this.object.scale.y=6;
       this.object.scale.z=6;
       */
-      this.object.add( gltf.scene);
+      this.object.add(gltf.scene);
     });
 
 
 
   }
 
-   return3DObject(){
+  return3DObject() {
     return this.object;
   }
 
-    update(time) {
-      //do nothing
-    }
+  update(time) {
+    //do nothing
+  }
 }
