--- public/js/charControls/BasicCharacterController.js
@@ -1,12 +1,0 @@
-class BasicCharacterController{
-    constructor(speed){
-        this.speed=speed;
-
-       
-    }
-
-    setMoveDistance(speed){
-        this.speed=speed;
-    }
-
-}
