+++ public/js/managers/EntityManager.js
@@ -1,20 +1,21 @@
-class EntityManager{
-    constructor(){
-        this.entities=[];
+//this class keeps track of our current entities
+export class EntityManager {
+    constructor() {
+        this.entities = [];
     }
 
     //typing the function outside will make it public as well (just like using this)
-    register(entity){
+    register(entity) {
         this.entities.push(entity);
     }
 
-    update(time){
-        for(let i=0;i<this.entities.length;i++){
+    update(time) {
+        for (let i = 0; i < this.entities.length; i++) {
             this.entities[i].update(time);
         }
     }
-    
-    clear(){
+
+    clear() {
         this.entities = [];
     }
